Change: 現在地取得等のgooglemap系の処理はstimulusを使用しない方向に変更
