+++ app/javascript/map/index.js
@@ -1,0 +1,1 @@
+import 'map/geocode'
