+++ app/views/search/departures/new.html.erb
@@ -45,3 +45,5 @@
     <div class="swiper-button-prev swiper-button-prev-2 !left-0.5 sm:!left-5 !text-cyan-600/80"></div>
   </div>
 </div>
+
+<%= javascript_include_tag Settings.google.map_url + 'callback=getCurrentLocation', defer: 'defer' %>
