+++ config/importmap.rb
@@ -5,6 +5,7 @@ pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
 pin "@hotwired/stimulus", to: "stimulus.min.js", preload: true
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js", preload: true
 pin_all_from "app/javascript/controllers", under: "controllers"
+pin_all_from "app/javascript/map", under: "map"
 pin "swiper", to: "https://ga.jspm.io/npm:swiper@8.4.2/swiper.esm.js"
 pin "dom7", to: "https://ga.jspm.io/npm:dom7@4.0.4/dom7.esm.js"
 pin "ssr-window", to: "https://ga.jspm.io/npm:ssr-window@4.0.2/ssr-window.esm.js"
