+++ config/settings/development.yml
@@ -1,5 +1,5 @@
 google:
-  map_url: 'https://maps.googleapis.com/maps/api/js?key=<%= Rails.application.credentials.google[:key] %>&callback=initMap'
+  map_url: 'https://maps.googleapis.com/maps/api/js?key=<%= Rails.application.credentials.google[:key] %>&'
   geocode_url: 'https://maps.googleapis.com/maps/api/geocode/json?key=<%= Rails.application.credentials.google[:key] %>&language=ja&region=JP&'
   nearby_url: 'https://maps.googleapis.com/maps/api/place/nearbysearch/json?key=<%= Rails.application.credentials.google[:key] %>&language=ja&'
 redis:
